Moved grades page into main and made necessary adjustments elsewhere
