+++ src/fe-app/src/app/grade-calc/grade-calc.component.html
@@ -29,7 +29,7 @@
   <h3>Final Grade: {{ calculateFinalGrade() | number:'1.2-2' }}%</h3>
 
   <div class = "btn-back">
-    <button routerLink="/grades" style = 'background-color: #40403e;'>Back to Grades</button>
+    <button routerLink="/main/grades" style = 'background-color: #40403e;'>Back to Main Page</button>
 
   </div>
 </main>
