+++ src/fe-app/src/app/grade-calc/grade-calc.component.ts
