+++ src/fe-app/src/app/main/grades/grades.component.html
@@ -24,14 +24,10 @@
     </tbody>
   </table>
   <button class="btn btn-primary mx-5">Add Grade</button>
-  <button class="btn btn-primary mx-5" routerLink="/main">Back to Main Page</button>
-
-
-<div class="icon" style="float: right; margin-right: 10px; margin-top: -30px;">
-    <a href="/grades/grade-calc">
-        <img src="images/calc.png" alt="grade calc icon"
-            style="width: 55px; height: auto;" />
-    </a>
-</div>
-
+  <div class="icon" style="float: right; margin-right: 10px; margin-top: -30px;">
+      <a href="/grades/grade-calc">
+          <img src="images/calc.png" alt="grade calc icon"
+              style="width: 55px; height: auto;" />
+      </a>
+  </div>
 </div>
