+++ src/fe-app/src/app/main/main.component.html
@@ -36,7 +36,7 @@
                         style="width: 80px; height: auto;"/></a>
                 </div>
                 <div class="col-md-auto">
-                    <a routerLink="/grades"><img src="images/gradesIcon.png" alt="grades icon" 
+                    <a routerLink="grades"><img src="images/gradesIcon.png" alt="grades icon" 
                         style="width: 75px; height: auto;"/></a>
                 </div>
                 <div class="col-md-auto ms-auto" (click)="openPopup('add-task')">
